requirements.txt
DBUtils
pymysql
psutil
pandas
matplotlib
scapy
scikit-learn
reportlab
openpyxl
shap
dpkt
python-nmap